+++ core/CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file. This projec
 
 ### Fixed
 
++ [#1831](https://github.com/luyadev/luya/issues/1831) Fixed issue with create url and complex composition patterns.
 + [#1826](https://github.com/luyadev/luya/issues/1826) Ensure arrayable implementation for Link objects.
 + [#1830](https://github.com/luyadev/luya/issues/1830) Lazyload widget asset registration issue fixed when used in nested context.